CHG: extract the important functions in utils as independent modules
