+++ sssd/core/layers/s4/hippo/utils.py
@@ -1,21 +1,17 @@
 import math
-from typing import Dict, Tuple, Union
+from typing import Tuple
 
 import numpy as np
-import opt_einsum
 import torch
 from einops import rearrange
 from opt_einsum import contract_expression
 from opt_einsum.contract import ContractExpression
-from scipy import special as ss
 
 from sssd.utils.logger import setup_logger
 
 LOGGER = setup_logger()
 
 Matrix = np.ndarray
-MeasureArgs = Dict[str, Union[float, int]]
-CONTRACT = opt_einsum.contract
 
 _c2r = torch.view_as_real
 _r2c = torch.view_as_complex
@@ -86,122 +82,6 @@ def embed_c2r(A: Matrix) -> Matrix:
     return rearrange(A_padded, "m x n y -> (m x) (n y)")
 
 
-class TransitionMatrix:
-    def __new__(
-        cls, measure: str, N: int, **measure_args: MeasureArgs
-    ) -> Tuple[Matrix, Matrix]:
-        """
-        Generate transition matrices for different measures.
-
-        Args:
-            measure (str): Type of measure.
-            N (int): Size of the transition matrices.
-            **measure_args: Additional arguments specific to each measure.
-
-        Returns:
-            Tuple[Matrix, Matrix]: Transition matrices A and B.
-
-        Raises:
-            NotImplementedError: If the specified measure is not implemented.
-            ValueError: If invalid measure arguments are provided.
-        """
-        if measure == "lagt":
-            A, B = cls._translated_laguerre(N, **measure_args)
-        elif measure == "glagt":
-            A, B = cls._generalized_laguerre(N, **measure_args)
-        elif measure == "legt":
-            A, B = cls._translated_legendre(N)
-        elif measure == "legs":
-            A, B = cls._scaled_legendre(N)
-        elif measure == "fourier":
-            A, B = cls._fourier(N)
-        elif measure == "random":
-            A, B = cls._random(N)
-        elif measure == "diagonal":
-            A, B = cls._diagonal(N)
-        else:
-            raise NotImplementedError(f"Measure '{measure}' is not implemented.")
-
-        return A, B
-
-    @staticmethod
-    def _translated_laguerre(
-        N: int, **measure_args: MeasureArgs
-    ) -> Tuple[Matrix, Matrix]:
-        """
-        Generate transition matrices for Laguerre (translated) measure.
-        """
-        beta = measure_args.get("beta", 1.0)
-        A = np.eye(N) / 2 - np.tril(np.ones((N, N)))
-        B = beta * np.ones((N, 1))
-        return A, B
-
-    def _generalized_laguerre(
-        N: int, **measure_args: MeasureArgs
-    ) -> Tuple[Matrix, Matrix]:
-        """
-        Generate transition matrices for Generalized Laguerre measure.
-        """
-        alpha = measure_args.get("alpha", 0.0)
-        beta = measure_args.get("beta", 0.01)
-        A = -np.eye(N) * (1 + beta) / 2 - np.tril(np.ones((N, N)), -1)
-        B = ss.binom(alpha + np.arange(N), np.arange(N))[:, None]
-        L = np.exp(
-            0.5 * (ss.gammaln(np.arange(N) + alpha + 1) - ss.gammaln(np.arange(N) + 1))
-        )
-        A = (1.0 / L[:, None]) * A * L[None, :]
-        B = (
-            (1.0 / L[:, None])
-            * B
-            * np.exp(-0.5 * ss.gammaln(1 - alpha))
-            * beta ** ((1 - alpha) / 2)
-        )
-        return A, B
-
-    @staticmethod
-    def _translated_legendre(N: int) -> Tuple[Matrix, Matrix]:
-        Q = np.arange(N, dtype=np.float64)
-        R = (2 * Q + 1) ** 0.5
-        j, i = np.meshgrid(Q, Q)
-        A = R[:, None] * np.where(i < j, (-1.0) ** (i - j), 1) * R[None, :]
-        B = R[:, None]
-        A = -A
-        return A, B
-
-    @staticmethod
-    def _scaled_legendre(N: int) -> Tuple[Matrix, Matrix]:
-        q = np.arange(N, dtype=np.float64)
-        col, row = np.meshgrid(q, q)
-        r = 2 * q + 1
-        M = -(np.where(row >= col, r, 0) - np.diag(q))
-        T = np.sqrt(np.diag(2 * q + 1))
-        A = T @ M @ np.linalg.inv(T)
-        B = np.diag(T)[:, None]
-        B = B.copy()
-        return A, B
-
-    @staticmethod
-    def _fourier(N: int) -> Tuple[Matrix, Matrix]:
-        freqs = np.arange(N // 2)
-        d = np.stack([freqs, np.zeros(N // 2)], axis=-1).reshape(-1)[:-1]
-        A = 2 * np.pi * (np.diag(d, 1) - np.diag(d, -1))
-        A = A - embed_c2r(np.ones((N // 2, N // 2)))
-        B = embed_c2r(np.ones((N // 2, 1)))[..., :1]
-        return A, B
-
-    @staticmethod
-    def _random(N: int) -> Tuple[Matrix, Matrix]:
-        A = np.random.randn(N, N) / N
-        B = np.random.randn(N, 1)
-        return A, B
-
-    @staticmethod
-    def _diagonal(N: int) -> Tuple[Matrix, Matrix]:
-        A = -np.diag(np.exp(np.random.randn(N)))
-        B = np.random.randn(N, 1)
-        return A, B
-
-
 def generate_low_rank_matrix(
     measure: str, N: int, rank: int = 1, dtype: torch.dtype = torch.float
 ) -> torch.Tensor:
@@ -249,67 +129,6 @@ def generate_low_rank_matrix(
     if rank > d:
         P = torch.cat([P, torch.zeros(rank - d, N, dtype=dtype)], dim=0)  # (rank, N)
     return P
-
-
-def normal_plus_low_rank(
-    measure: str,
-    matrix_size: int,
-    correction_rank: int = 1,
-    dtype: Union[type(torch.float), type(torch.cfloat)] = torch.float,
-) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Normal plus Low-rank
-    Return eigenvalues, correction_matrix, transformed_vector, unitary_matrix such that
-    (eigenvalues - correction_matrix correction_matrix^*, transformed_vector) is unitarily equivalent
-    to the original HiPPO A, B by the matrix unitary_matrix.
-    i.e.,
-       1. A = unitary_matrix[eigenvalues - correction_matrix correction_matrix^*]unitary_matrix^*
-       2. B = unitary_matrix B
-
-    Args:
-    - measure: The type of measure used for the HiPPO matrix.
-    - matrix_size: The size of the HiPPO matrix.
-    - correction_rank: The rank for the correction matrix P.
-    - dtype: The data type for the tensors.
-
-    Returns:
-    - w: The eigenvalues of the matrix AP.
-    - P: The rank correction matrix.
-    - B: The transformed input vector.
-    """
-    assert dtype in (
-        torch.float,
-        torch.cfloat,
-    ), "dtype must be torch.float or torch.cfloat"
-    half_size = matrix_size // 2
-    if measure == "random":
-        dtype = torch.cfloat if dtype == torch.float else torch.cdouble
-        w = -torch.exp(torch.randn(half_size, dtype=dtype)) + 1j * torch.randn(
-            matrix_size // 2, dtype=dtype
-        )
-        P = torch.randn(correction_rank, half_size, dtype=dtype)
-        B = torch.randn(half_size, dtype=dtype)
-        return w, P, B
-
-    A, B = TransitionMatrix(measure, matrix_size)
-    A = torch.as_tensor(A, dtype=dtype)
-    B = torch.as_tensor(B, dtype=dtype)[:, 0]
-
-    P = generate_low_rank_matrix(
-        measure, matrix_size, rank=correction_rank, dtype=dtype
-    )
-    AP = A + torch.einsum("...i,...j->...ij", P, P.conj()).sum(dim=-3)
-    w, V = torch.linalg.eig(AP)
-
-    # Keep only one of each pair of complex conjugate eigenvalues
-    mask = w.imag >= 0
-    w = w[mask]
-    V = V[:, mask]
-
-    V_inv = V.conj().transpose(-1, -2)
-    B = CONTRACT("ij, j -> i", V_inv, B.to(V.dtype))  # V^* B
-    P = CONTRACT("ij, ...j -> ...i", V_inv, P.to(V.dtype))  # V^* P
-
-    return w, P, B
 
 
 def cauchy_cpu(v: torch.Tensor, z: torch.Tensor, w: torch.Tensor) -> torch.Tensor:
