+++ tests/core/hippo/test_hippo_utils.py
@@ -3,7 +3,6 @@ import pytest
 import torch
 
 from sssd.core.layers.s4.hippo.utils import (
-    TransitionMatrix,
     cauchy_cpu,
     cauchy_wrapper,
     compute_fft_transform,
@@ -16,7 +15,6 @@ from sssd.core.layers.s4.hippo.utils import (
     get_output_contraction,
     hurwitz_transformation,
     low_rank_woodbury_correction,
-    normal_plus_low_rank,
     power,
 )
 
@@ -37,66 +35,6 @@ def test_power_without_v():
     )
     result = power(L, A)
     assert torch.allclose(result, expected_result)
-
-
-def test_transition_lagt():
-    N = 10
-    beta = 1.5
-    A, B = TransitionMatrix("lagt", N, beta=beta)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-    assert np.allclose(A, np.eye(N) / 2 - np.tril(np.ones((N, N))))
-    assert np.allclose(B, beta * np.ones((N, 1)))
-
-
-def test_transition_glagt():
-    N = 10
-    alpha = 0.5
-    beta = 0.01
-    A, B = TransitionMatrix("glagt", N, alpha=alpha, beta=beta)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_legt():
-    N = 10
-    A, B = TransitionMatrix("legt", N)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_legs():
-    N = 10
-    A, B = TransitionMatrix("legs", N)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_fourier():
-    N = 10
-    A, B = TransitionMatrix("fourier", N)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_random():
-    N = 10
-    A, B = TransitionMatrix("random", N)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_diagonal():
-    N = 10
-    A, B = TransitionMatrix("diagonal", N)
-    assert A.shape == (N, N)
-    assert B.shape == (N, 1)
-
-
-def test_transition_invalid_measure():
-    N = 10
-    with pytest.raises(NotImplementedError):
-        TransitionMatrix("invalid_measure", N)
 
 
 def test_embed_c2r():
@@ -183,65 +121,6 @@ def test_rank_correction_invalid_measure():
 def test_rank_correction_dtype():
     P = generate_low_rank_matrix("legs", 5, 1, dtype=torch.float64)
     assert P.dtype == torch.float64
-
-
-# nplr
-@pytest.fixture
-def setup_data():
-    measure = "random"
-    N = 10
-    rank = 1
-    dtype = torch.float
-    return measure, N, rank, dtype
-
-
-def test_nplr_shapes_and_types(setup_data):
-    measure, N, rank, dtype = setup_data
-    w, P, B = normal_plus_low_rank(measure, N, rank, dtype)
-
-    # Check if the shapes are correct
-    assert w.shape == (N // 2,)
-    assert P.shape == (rank, N // 2)
-    assert B.shape == (N // 2,)
-
-    # Check if the types are correct
-    assert w.dtype == (torch.cfloat if dtype == torch.float else torch.cdouble)
-    assert P.dtype == w.dtype
-    assert B.dtype == w.dtype
-
-
-def test_nplr_values(setup_data):
-    torch.manual_seed(1)
-    measure, N, rank, dtype = setup_data
-    w, P, B = normal_plus_low_rank(measure, N, rank, dtype)
-
-    # Check if the imaginary parts of w are within the range [-2, 2]
-    assert torch.all(
-        (w.imag >= -2) & (w.imag <= 2)
-    ), "Imaginary part of w should be within [-3, 3]"
-    assert torch.all(torch.abs(P) <= 5), "Elements of P should be within [-5, 5]"
-
-
-# Test to ensure that the eigenvalues 'w' are complex numbers
-def test_nplr_w_complex(setup_data):
-    _, N, _, dtype = setup_data
-    w, _, _ = normal_plus_low_rank("random", N, dtype=dtype)
-    assert all(
-        torch.is_complex(val) for val in w
-    ), "Eigenvalues w should be complex numbers"
-
-
-# Test to ensure that 'P' has the correct rank
-def test_nplr_p_rank(setup_data):
-    _, N, rank, dtype = setup_data
-    _, P, _ = normal_plus_low_rank("random", N, correction_rank=rank, dtype=dtype)
-    assert P.shape[0] == rank, f"Matrix P should have rank {rank}"
-
-
-# Test to check if 'nplr' function handles invalid 'measure' input
-def test_nplr_invalid_measure():
-    with pytest.raises(NotImplementedError):
-        normal_plus_low_rank("invalid_measure", 10)
 
 
 def test_cauchy_cpu():
